refactor: extract measureTreeMaxRowWidth to AutoFitCalculator
- Move tree row measurement to shared
  AutoFitCalculator.measureTreeMaxRowWidth
- ToolWindowAutoFitter delegates instead
  of inline loop
- GitPanelAutoFitManager delegates to same

Impact: single source of truth for tree
width measurement across all auto-fitters
